Appled patch from Matthias Kretz, with an additionnal kdWarning, to prevent crashes when killing a job (the patch, not the kdWarning) :)
svn path=/trunk/kdelibs/; revision=74262

+++ kio/observer.cpp
@@ -82,7 +82,12 @@ void Observer::jobFinished( int progressId )
 void Observer::killJob( int progressId )
 {
     KIO::Job * job = m_dctJobs[ progressId ];
-    assert(job);
+    if (!job)
+    {
+        kdWarning() << "Can't find job to kill ! There is no job with progressId=" << progressId << " in this process" << endl;
+        m_dctJobs.remove( progressId );
+        return;
+    }
     job->kill( false /* not quietly */ );
 }
